@@ -1,44 +1,41 @@
 {
   "name": "suttisak-personal-site",
-  "private": true,
+  "packageManager": "yarn@4.9.2",
   "description": "Suttisak Profile and Portfolio site",
   "version": "3.0.0",
   "author": "Suttisak Denduangchai <suttisak.lukesp@outlook.com>",
   "dependencies": {
-    "@mdx-js/react": "^2.1.5",
-    "gatsby": "^5.0.1",
+    "@mdx-js/react": "^3.1.0",
+    "gatsby": "^5.14.5",
     "gatsby-plugin-anchor-links": "^1.2.1",
-    "gatsby-plugin-canonical-urls": "^5.0.0",
-    "gatsby-plugin-google-gtag": "^5.0.0",
-    "gatsby-plugin-image": "^3.0.0",
-    "gatsby-plugin-manifest": "^5.0.0",
-    "gatsby-plugin-mdx": "5.0.0",
-    "gatsby-plugin-no-sourcemaps": "^5.0.0",
-    "gatsby-plugin-react-helmet": "^6.0.0",
-    "gatsby-plugin-sass": "^6.0.0",
-    "gatsby-plugin-sharp": "^5.0.0",
-    "gatsby-plugin-sitemap": "^6.0.0",
-    "gatsby-plugin-typescript": "^5.0.0",
-    "gatsby-source-filesystem": "^5.0.0",
-    "gatsby-transformer-json": "^5.0.0",
-    "gatsby-transformer-sharp": "^5.0.0",
-    "gh-pages": "^4.0.0",
+    "gatsby-plugin-canonical-urls": "^5.14.0",
+    "gatsby-plugin-google-gtag": "^5.14.0",
+    "gatsby-plugin-image": "^3.14.0",
+    "gatsby-plugin-manifest": "^5.14.0",
+    "gatsby-plugin-mdx": "5.14.1",
+    "gatsby-plugin-no-sourcemaps": "^5.14.0",
+    "gatsby-plugin-react-helmet": "^6.14.0",
+    "gatsby-plugin-sass": "^6.14.0",
+    "gatsby-plugin-sharp": "^5.14.0",
+    "gatsby-plugin-sitemap": "^6.14.0",
+    "gatsby-plugin-typescript": "^5.14.0",
+    "gatsby-source-filesystem": "^5.14.0",
+    "gatsby-transformer-json": "^5.14.0",
+    "gatsby-transformer-sharp": "^5.14.0",
+    "gh-pages": "^6.3.0",
     "jsxgraph-react-js": "^1.0.4",
-    "katex": "^0.16.3",
-    "moment": "^2.29.4",
-    "prism-react-renderer": "^1.3.5",
-    "prismjs": "^1.29.0",
+    "katex": "^0.16.22",
+    "moment": "^2.30.1",
+    "prism-react-renderer": "^2.4.1",
+    "prismjs": "^1.30.0",
     "prop-types": "^15.8.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
-    "react-intersection-observer": "^9.4.0",
+    "react-intersection-observer": "^9.16.0",
     "rehype-katex": "^6.0.2",
-    "rehype-slug": "^5.1.0",
-    "remark-math": "^5.1.1",
-    "remark-slug": "^7.0.1",
-    "remark-toc": "^8.0.1",
-    "sass": "^1.56.1"
+    "rehype-slug": "^6.0.0",
+    "sass": "^1.89.2"
   },
   "keywords": [
     "gatsby"
@@ -56,10 +53,10 @@
     "test": "echo \"Write tests! -> https://gatsby.app/unit-testing\""
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.13",
-    "gatsby-plugin-postcss": "^6.0.0",
-    "postcss": "^8.4.19",
-    "tailwindcss": "^3.2.4"
-  },
-  "packageManager": "yarn@3.2.4"
+    "@types/prismjs": "^1",
+    "autoprefixer": "^10.4.21",
+    "gatsby-plugin-postcss": "^6.14.0",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^3.4.17"
+  }
 }
