@@ -1,6 +1,4 @@
-module.exports = {
-  plugins: {
+export const plugins = {
     tailwindcss: {},
     autoprefixer: {},
-  },
 }
